why is it called stealthed like what also fixed those annoying warnings

+++ Amalgam/src/Features/Misc/AutoQueue/AutoQueue.cpp
@@ -343,7 +343,7 @@ void CAutoQueue::SearchCommunityServers()
 	m_hServerListRequest = I::SteamMatchmakingServers->RequestInternetServerList(
 		440,
 		&pFilters,
-		vFilters.size(),
+		uint32(vFilters.size()),
 		this
 	);
